kehadiran: make policy to  resource

+++ app/Http/Controllers/KehadiranController.php
@@ -5,10 +5,11 @@ namespace App\Http\Controllers;
 use App\Http\Requests\Kehadiran\StoreRequest;
 use App\Http\Requests\Kehadiran\UpdateRequest;
 use App\Http\Resources\KehadiranResource;
+use App\Models\Kehadiran;
 use App\Services\KehadiranService;
 use App\Traits\ApiResponse;
-
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class KehadiranController extends Controller
 {
@@ -39,30 +40,40 @@ class KehadiranController extends Controller
 
     public function index(Request $request)
     {
+        Gate::authorize('viewAny', Kehadiran::class);
+
         $data = $this->kehadiranService->getAllKehadiran($request);
         return $this->success('Berhasil mengambil semua data kehadiran', KehadiranResource::collection($data));
     }
 
     public function show($id)
     {
+        Gate::authorize('view', Kehadiran::class);
+
         $data = $this->kehadiranService->getKehadiranById($id);
         return $this->success('Berhasil mengambil data kehadiran', new KehadiranResource($data));
     }
 
     public function store(StoreRequest $request)
     {
+        Gate::authorize('create', Kehadiran::class);
+
         $data = $this->kehadiranService->createKehadiran($request->all());
         return $this->success('Berhasil menambahkan data kehadiran', new KehadiranResource($data), 201);
     }
 
     public function update(UpdateRequest $request, $id)
     {
+        Gate::authorize('update', Kehadiran::class);
+
         $data = $this->kehadiranService->updateKehadiranById($request->all(), $id);
         return $this->success('Berhasil mengubah data kehadiran', new KehadiranResource($data));
     }
 
     public function destroy($id)
     {
+        Gate::authorize('delete', Kehadiran::class);
+
         $this->kehadiranService->deleteKehadiranById($id);
         return $this->success('Berhasil menghapus data kehadiran');
     }
